requirements.txt
google-genai>=0.3.0
playwright>=1.42.0
pydantic>=2.6.0
gspread>=6.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
